Define git-lfs in the plugins repo under actions (#15)
Define git-lfs in the plugins actions repo. 

They'll be double defined for a bit - but namespace resolution would kick in for now until we have hoisting and then it'll be a conflict that needs resolution. In the meantime the docs point to this as an example so thought it should be in there.