Improve testing workflow (#68)
Changes targeted by this PR:
- Avoid snapshot assertions failing when it's a new version that doesn't
have a snapshot. This includes removing the requirement to specify
output files, and now snapshots are always autogenerated. See comment in
`linterCheckTest()` for specifics on this design. This decision was
rooted in a couple factors, namely:
- As identified in #65, the lack of a snapshot for a specific version
was causing unnecessary failures.
- Linters can still change their messages, outputs, etc (as sqlfluff did
yesterday). This is something that we don't want to cause too much
friction, but we still want to call it out as a noticeable change. If
this happens during someone's workflow (should still be relatively
infrequent), now simply run the specific command in tests/readme.md and
validate expected output. _We may want to highlight this specific
command better or with more visibility._
- The JSON assertions and snapshot code were a bit redundant and
unnecessarily complex. This should simplify and be flexible enough for
the near future. If we want any additional assertions, we can do so via
the `postCheck` hook.
- Make tests run in parallel. This also simplifies the setup for the
`config_check_test.ts`. It's a tad redundant, but its presence makes
triaging problems a bit easier.
- Attempt to reduce flakiness (the parallelism appears to reduce
flakiness to about 1-3%--it'd be nice to have analytics on this). I'll
continue investigating alternatives for dealing with the daemon
lifetime.

Average test time right now locally with these 4 tests is about 10
seconds in total :)
They're about a minute in total on CI because of the need to download
and install linters.

Remaining todos (unordered):
- Investigate daemon lifetime changes
- Modify sandboxing to shadow tree from root
- Use debugging library in testing
- Add contributing.md
- Add option for specifying a range of linter versions (ideally for
nightlies)
- Migrate rest of definitions