Improve testing workflow further, add debug and coverage (#69)
This is a followup to #68 and introduces the following:
1. Added debug library (instructions for use in `tests/readme.md`).
Namespaces are of form: `${Location}/${linter}/${#}`.
2. Add test that asserts linter test coverage is required.
3. Add `Snapshots` option to `PLUGINS_TEST_LINTER_VERSION` that runs
tests on each snapshot version that is provided (eventually we can run
this on nightlies, but for now it's a useful tool for testing new cli
versions).

Also addressed are:
1. Linters and formatters now have unified behavior when it comes to
checking snapshot versions. They may not necessarily provide the same
version set, however, since formatters tend to be more stable.
2. Modified test directory structure to use a `test_data` folder and no
`__snapshots__` folder. See the modifications in `tests/readme.md`.
3. Added information about snapshot resolution to `tests/readme.md`.
4. Rename `PLUGINS_TEST_NEW_SNAPSHOT` to
`PLUGINS_TEST_UPDATE_SNAPSHOTS`.
5. Rename snapshot files to syntax:
`${linter}_v${linterVersion}_${prefix}.${check/fmt}.shot`.
6. Added comments for clarity throughout.