Cache downloads on PR runs (#72)
Accomplishes two things:
1. Cache linter downloads across PR runs
2. Fixes a small bug with how preCheck and postCheck callbacks are
invoked (this is an important fix for getting git-diff-check working).

When running 20+ tests, this brought the PR time from 10+ minutes down
to 4 minutes. I still have some timeouts to investigate on large numbers
of tests in CI.

I don't think there's a real need to flush/reset these downloads, but
when we add nightly workflows we can implement that.