Testing improvements to support more linters (#78)
Accomplishes the following:
- Streamlines preCheck hooks by exposing more helper methods in the
TrunkDriver for file manipulation.
- Adds more custom helper methods for running `check` and `fmt` without
specifying specific targets. This should enable most of the remaining
linters. The thinking behind this setup is as follows:
- Most simple use cases should use `linterCheckTest` or `linterFmtTest`,
which should supply a decent amount of isolation and clarity on test
inputs and outputs.
- For most other use cases, such as when targets != inputs (e.g. clippy,
or trigger-based linters), or when specific functionality is being
tested, arguments can be specified at will using `customLinterCheckTest`
or `customLinterFmtTest`.
- Fixes a small bug where multiple launchers invocations could collide
over non-atomic operations. This helps with the flakiness, but I still
haven't tracked down the root cause.
- Adds a test for sort-package-json, demonstrating these custom helper
methods (note that a custom test is necessary in this case since
`package.json` can never be `package.in.json`.

I will follow this up with additional linter migrations. And also fix
the flaky test timeout.