Fix the plugin version used in the upload (#152)
Previously, this was hard-coded. Now, use the version that the tests
actually run against. Thanks to the checkout dance, this will take
effect on the next nightly, even if we don't make a release.

Remaining work includes potential slack notifications, which would
probably require the use of a separate failure.json file (this could
include information, such as `MISMATCHED`).