Fix flakiness (#207)
Attempts to fix the following:
- sqlfluff: adds a new snapshot for the 2.0.0 release, and add logic to
LUV test runners to copy over old snapshots that have been marked with
the `// RELEASE` string (can be safely removed after next release)
- We don't always want the snapshot to be available when testing against
the latest release, since the snapshot could be related to new testing
logic, which might not appear in the release.
- We do, however, want to have the ability to copy it over, such as in
the current case when there is no problem with the new sqlfluff other
than it requires new snapshots.
- eslint: increases timeout
- cspell: I'm disabling this now to unblock the other PRs but am still
investigating. So far, I've determined that cspell is more flaky on Mac
when all tests are run, indicating some performance-related condition or
race.