Improve versioning and installs for commitlint (#196)
Fixes commitlint to at least have a default config specification
(originally it errored, complaining about the lack of one).

Also provide instructions on how to override the listed commitlint
plugins (the recommendation requires >1.6.0, hence the specification. I
can make this >=1.6.0 if desired).

In the future, we will probably want to explicitly version these sort of
actions via config so that we can guarantee hermetic treatment.