Avoid inconsistent diagnostics from cspell (#209)
The original issue was that sometimes cspell tests were flaky,
particularly on the GH Mac runners (e.g.
[nightly](https://github.com/trunk-io/plugins/actions/runs/4465976153/jobs/7843594981)).
I was unable to reproduce under simulated conditions, but I was able to
get similar results locally by shrinking the `suggestionsTimeout`
setting down to 5 ms. This suggested that due to resource constraints,
the Mac tests were timing out and terminating early with a smaller set
of cspell suggestions.

While we won't know for sure for at least another release if this is the
root cause of the issue, adding a default config with a
`suggestionsTimeout` of 5000 ms can hopefully prevent any users from
hitting this issue.