Improve config_check_test (#220)
When we added relative path support for sourcing plugins (and really
even a bit before that), config_check_test was redundant, in that all of
the linter tests would fail anyway if a config was bad. There wasn’t a
need for a separate test.

With that being said, we do some manual tests with the
`required_trunk_version` that could be automated. Thus, I essentially
had 2 options:
1. Change the linter tests, instead of sourcing the entire repo’s
plugin.yaml’s, only source the filetypes and the linter in question.
This would be a slight performance improvement as well as some greater
test isolation, but isn’t really necessary. I prefer the fail fast as it
is now.
2. Change the config_check_test to run on `required_trunk_version` .
Because of the semver `>`/`>=`, it’s simpler to just include a cli
version to run in the tests, as I’ve done here (rather than inferring).
I've added `trunk-toolbox` to support this.