Add stylua (#222)
Adds [stylua](https://github.com/JohnnyMorganz/StyLua/tree/main), an
opinionated formatter (and a bit more) for lua files. A couple notes:
1. Adds the comment format and file type for lua, since this is our
first lua tool.
2.
[`--search-parent-directories`](https://github.com/JohnnyMorganz/StyLua/tree/main#finding-the-configuration)
is required in the run command in order for a root level (or other)
config file to be respected. However, this falls back to checking
`XDG_CACHE_HOME` for a config. This can violate hermeticity across a
repo, so I've chosen to dump the default config whenever someone runs
`trunk check enable stylua`. This should prevent the likelihood of these
cases happening.

Thanks @docwhat for suggesting and starting this!