Fix missing default_*.yaml values (#231)
Address any fixes that were included in the `default_*.yaml` files that
hadn't gotten ported over. I changed the implicit plugin ref to HEAD on
this branch and got most of the existing tests to pass after porting
these changes.

We could just delete the deprecated linters (it's been a while) and
modify our tests, but I think it's fine to keep them like this for now.
This PR deprecates `terraform-validate` for the first time (had
previously been a TODO to deprecate it).

This should be the last change necessary before getting a release out.