Added trufflehog (#241)
Added trufflehog as a security linter.

One thing of note is that it only reports an AWS secret if it finds both
the access key ID and the secret access key. This means that it would
find no errors in [the test file for
gitleaks](https://github.com/trunk-io/plugins/blob/main/linters/gitleaks/test_data/basic.py).