Add testing for symlinking configs and audit failures (#254)
#250 identified a blindspot in our test coverage. Config files should
not be included in the testing sandbox creation process. They should be
dumped into `.trunk/configs` during the linter enable process. This
change and investigation spawned a couple of insights:
- gokart doesn't work for shadow trees. We need to follow-up with an
option to expand the shadow tree to only symlink individual files.
- cspell doesn't respect a config file if it is a symlink in the context
of other symlinked files. We need to remove `in_place: true` and copy it
directly.
- ruff-nbqa output format changed slightly, but it's okay since it's a
pass_fail linter.
- fix some existing clang-tidy tests