Fix custom fmt test snapshots (#256)
`customLinterFmtTest`s were written originally to only have one path to
snapshot, and they didn't interact nicely with other format tests. Fix
the regex and the semantics to support multiple snapshots. This was
breaking [`main`
tests](https://github.com/trunk-io/plugins/actions/runs/4730053093/jobs/8404731644)
and is a plugins release blocker.

If need be, it should be straightforward to refine this in the future
(as well as for `customLinterCheckTest`).