Add sandbox-type support (#258)
Previously, we were using `in_place: true` as a hacky mechanism for
copying files into a sandbox to be linted. `in_place` is meant to be
reserved for formatters. Now, we have the `sandbox_type: normal |
expanded | copy_targets` flag to replace it, as well as to enable
`gokart` to correctly run on directories with an expanded shadow tree
(in_place does not work for linters that run on directory targets).

Will require a `required_trunk_version` bump.