Add suggest-type support (#259)
Whether a linter was auto-enabled/auto-recommended or not used to be
based on the complex interaction of several fields. Namely, in order to
be auto-enabled without prior existence of a config, a linter had to
have:
1. is_recommended: true (default true)
2. good_without_config: true OR have a recommended direct_config to dump

Now, this has been simplified to use the `suggest_if` field (options:
`never` | `config_present` | `files_present` | `legacy`). Uses the new
syntax for all plugin linters and fixes regression for several linters.
See comment about new behavior enumerated by linter.

Also modified testing to be more faithful to the actual behavior.