include suggest_if and fix* buf breaking (#272)
Improves buf-breaking and reenables its tests in order to run closer to
how it was intended to be. Some notes:
- Buf breaking will error out and show some diagnostics early, so you
have to fix those and rerun in order to see new diagnostics
- Have to disable following symlinks to avoid seeing test data in our
own plugins repo symlink :)
- Cannot cache results because of git awareness
- Unfortunately, we do still need to have upstream turned on in order
for this to work "properly"
- For some results, buf breaking non-deterministically favors one
ordering over another. See test note

**Because of this, reviewing this PR probably necessitates a brief
discussion about expected behavior for it**

Things I tested:
- Works with adding new files
- Works with modifying existing files, including in different git setups
(on main, not on main)
- When proto files are deleted or moved, doesn't work great, but doesn't
fail. This is kind of expected, and should be resolved with trunk skip
check.

Also adds in some overdue `suggest_if`s