Partial terragrunt integration (#277)
- add .envrc to plugins
- terraform tool integration
- integrating terragrunt fmt

I would also like to integrate terragrunt's validation tool - as
mentioned in standup, it requires terraform to be in your `PATH`.
However, even with `terraform` in path, I cannot get it to work
properly. There is another environment variable, `TERRAGRUNT_TFPATH`,
which allows me to get this to work, but we'd have to evaluate that to
point to the `$(which terraform)`. So, punting on that integration for
now.