Bump min version for 5 go linters (#282)
The following go linters are not able to query latest (using `trunk
check enable --bypass-validated <linter>`) with our current setup:
- cue-fmt
- gofmt (uses go download)
- gofumpt
- goimports
- shfmt

I've bumped their min versions and made notes about this limitation. To
fix this more sustainably (easy for 3 of these, difficult for remaining
2), we'll have to add new functionality to the cli for
`--bypass-validated`, such as something like:

```yaml
lint:
  definitions:
    - name: cue-fmt
       ...
       runtime: go
       package: cuelang.org/go/cmd/cue
       query_url: github.com/cue-lang/cue
```

The reason we can't just override package today is because we get the
install error:
```
    +     stderr: |
    +       go: downloading github.com/cue-lang/cue v0.4.3
    +       go: github.com/cue-lang/cue@v0.4.3: github.com/cue-lang/cue@v0.4.3: parsing go.mod:
    +           module declares its path as: cuelang.org/go
    +                   but was required as: github.com/cue-lang/cue
```

Open to thoughts. Not prioritizing this for now.