Add fuzzy test matching for non-idempotent linters (#313)
Adds a fuzzy check test that operates as follows:
- Provide most of the same arguments as other tests
- The test strips the snapshot of its `issues` so that it's only an
empty list
- Create an `expected_issues.json` file with the expected issues
- Call `fuzzyLinterCheckTest` and pass in a fuzzyMatcher, which includes
the expected issues and a minimum required threshold for overlap

The current overlap for osv-scanner is 30 of 51, and the current overlap
for trivy is 12 of 16. Using the custom matcher, it still prints all of
the output issues, and it still does the normal assertion behavior for
the rest of the landing state. If the test fails due to a significant
drift in fileIssues, we can either adjust the threshold or modify
`expected_issues.json` to include a more updated list. I estimate that
upkeep for these non-idempotent linters will be reduced to once very few
months, barring other breaking changes.

This implementation leaves room for version-specific expected issues
still, allowing us to preserve the existing snapshot logic as needed.
Regardless, this shouldn't be necessary in most cases since these
linters tend to only have one versioned snapshot.