Fix nightly workflow to use workflow's CLI version (#376)
Fix the workflow to use the cli version on main/workflow branch rather
than the cli version and trunk.yaml from the release branch. This wasn't
a problem before, but because we need our updated querying logic for the
pypi fix, we now require these to match. My approach here is perhaps a
bit verbose, but it gets the job done.

Successful
[run](https://github.com/trunk-io/plugins/actions/runs/5605161418/jobs/10253982794)
(yamllint is now correctly versioned).