Setup windows testing and prep for Windows (#330)
Final major PR to get plugins repo working for Windows:
- Mark additional linters as unsupported as applicable.
- Add a new workflow, `windows_nightly.yaml`, which runs Windows tests
on linters and tools. Windows is slower and less stable, so we should
not gate on this. As needed we can turn it on for individual PRs or run
it prior to releases. Eventually we can setup Slack notifications for
it. Tests run using the powershell launcher. We can continue to burndown
this list as time goes on.
- Includes various fixes for linters in order to reach the ~74%
threshold.
- Improved logs/debugging for some CLI execution. Also includes
try/catch for some (currently) expected Windows CLI deficiencies.

Most recent complete
[run](https://github.com/trunk-io/plugins/actions/runs/5613908757/job/15210942208?pr=330)

We will likely have 1-2 final small follow-ups.