Fix osv-scanner line numbers (#384)
The previous logic was reading lines of a lockfile line-by-line inside
the loop that iterated issues, but as the issue iteration marched on,
the read position in the lockfile did too (didn't ever reset via
fseek(0)), so what ended up happening was some issues would have line
numbers but most wouldn't. This changes it to just read the entire
lockfile into a string, and then iterate the lines of that string
multiple times.