Support querying latest on Windows (#415)
We need to have the runtime in the PATH in order to query latest. On
most of the Linux setups, we inherit the system PATH intentionally for
these queries, which usually resolves npm/node. But on Windows (and for
hermeticity), we need to use our managed runtimes to query. Setup every
test with these runtimes enabled and downloaded so that we can query.

This is a bandaid but should be sufficient since in most cases this
command is invoked with runtimes already enabled and/or just for the
plugins testing.

Also fixes cspell tests on Windows by modifying the
`testCreationFilter`.