Convert cspell tests to use fuzzy matching (#493)
Cspell updated their dictionaries twice this week, and I'd rather use a
fuzzy matcher than keep updating snapshots. There are 2 core changes
here:
- Normalize the cspell diagnostics in the tests to always output
suggestions in alphabetical order (they're normally reported in
recommended order, but we don't care much about that for the sake of the
tests)
- Use a fuzzy matcher. That way, if suddenly a new suggestion appears or
the number of suggestions changes, we won't fail the test until a number
of them accumulate (so we can maintain this at a lower frequency)

I've also bumped the KGV to 7.0.0 for ease of testing

In the event that this test fails in the future, depending on the result
there will be 3 possible approaches:
- If we want to maintain our older snapshots, we can modify
`fileIssueAssertionCallback` to take in the version of the linter
running as well so we can triage to a relevant JSON file.
- If there's just a couple mismatched diagnostics, we can either update
the expected issues JSON files or reduce the required overlap threshold
- Another approach. Fortunately, this is fairly extendable.

As an aside, the fuzzy matcher experience is pretty good. When it fails,
the matcher outputs the issues that we received, so it's easy to debug.
The only tough part is generating the JSON file in the first place. If
need be, I'll write an action or script for this in the future.