Fix parsers for non-utf8 filepaths (#494)
On Windows, we were crashing because our custom parsers were outputting
incorrect non-utf8 filepaths with bad encoding. It's unclear why this
was only happening on Windows, but this fix helps for all of our linters
with custom parsers.

We will also improve our discard logic to not crash when we get a bad
path like that.