Make Windows Green (#510)
Overview of changes:
- Bump timeouts for Windows. We thought that some of these were hanging
so we killed them early, but actually they're just slow
- Tracked down the eslint issue. When run in parallel, certain node exec
calls will create circular JSON errors when an error occurs (usually
ENOENT or timeout), leading to dubious and unclear error messages. I
modified this to be a tool-based setup and use spawn (the underlying
callee).
- Couldn't really find the root cause of the golangci-lint issue. I
ended up turning some of the tests off and here are some of my other
findings:
- typecheck precedes all else, especially on Windows. And for some
reason (likely differently-built system libs), the typecheck output
shows up in errors instead of issues on Windows, leading to the original
snapshot issues.
- Even though golangci-lint correctly realizes that its sub-linters are
enabled, it doesn't report the diagnostics we expect on Windows. I
timeboxed this and came up empty. If someone stumbles across an
explanation at some point, we will likely want to modify our
Windows-specific setup for golangci-lint.

Windows should now be
[green](https://github.com/trunk-io/plugins/actions/runs/6480661270/job/17596621934)
(passing except for the changes in #515)