Revert trufflehog off-by-1 change (#536)
Reverts changes from #532, since 3.60.3 is out and reverts the change.
Marked 3.60.1 and 3.60.2 as `known_bad_versions` and removed the
unnecessary snapshots. Note that the off-by-one discrepancy between git
and filesystem outputs is still present, but we handle this through our
custom parser. If they fix this in the future we will have to modify our
parser to take in inputs, like #532.