fix(trufflehog): report the correct line numbers (#531)
Trufflehog can report wildly incorrect line numbers for secrets, and
even report multiple incorrect issues for a given secret. This PR
changes our trufflehog parser to scan the file for the secret and report
the correct line number. As the trufflehog can report duplicates, we
only output a lint issue once for a given secret.

This is tested with `wrong_line_number.in.ts`, which trufflehog reports
4 issues on incorrect lines for 2 identical secrets. The parser finds
these two secrets on their separate lines.

This only fixes the issue for trufflehog filesystem - I'm not sure if
this an issue in trufflehog git, and the files may be changed/deleted
since the past.