Upgrade CLI and handle eslint batching (#585)
Upgrade CLI and clear release snapshot marks.

Recent changes to retry batching cause eslint to output several failures
for one of the test instead of just 1. We will fix this in a future
version and update the snapshot and mark for release.

Also note that this upgrade should resolve go install issues.