Auto generate snapshots from nightly (#596)
Attach additional information about the type of testing failure that
occurred (i.e. `passed`, `unknown`, `task_failure`, or
`assertion_failure`). If the latter occurred on all platforms (other
than skipped), we can likely conclude that either:
1. The linter released a new version with a differing output
2. The linter's output relies on a live service, and all of its
snapshots will need to be changed

The second case is out of scope for now, so for case 1 we will
proactively generate a snapshot, mark for release, and create a PR. If
this PR passes tests and a brief sanity check, then we will be back to
business as usual.

Notes:
- Because it is marked for release, in most cases we won't have to worry
about this being spammy for other edge cases. If issues do occur though,
we can always add an ignore filter to the jest rerun invocation.
- The failure notification will include _`(rerunning)`_ if it triggers a
rerun test.
- If something goes wrong or a snapshot cannot be generated on the
rerun, the job will fail, and it can be triaged through other nightly
email notification triage.
- If we ever add any additional complexity to the `parse/index.ts`
script, we should probably write a test for it.

Successful test
[run](https://github.com/trunk-io/plugins/actions/runs/7380972339/job/20079163879)
creating [PR](https://github.com/trunk-io/plugins/pull/604/files)