Enable more batching (#591)
We used to be conservative about when to enable batching due to the
worse error messages when a linter failed, but with the new batch
bisection feature, we can enable batching in many more places. Even for
very fast linters like clang-format, batching is a large speedup. The
only place we would not want to batch is for very expensive linters like
clang-tidy or if unsupported by the linter's CLI. I tested each linter
with a manual batch.

This enables batching on all formatters and also checkov.