Fix snapshot auto-generation bug (#619)
Fixes another bug in snapshot generation that led to the creation of
https://github.com/trunk-io/plugins/pull/618. Depending on the order of
the test parsing, mismatched versions were being treated as a "pass",
which was an overloaded state and led to snapshot auto-generation. I
added the `Skipped` failure enum type, which makes these states more
explicit and fixes the underlying issue.

I also tested manually that this still has the fix from #612.