Add `commitizen` action/tool (#660)
Hello!
This PR adds https://github.com/commitizen/cz-cli as both an action and
a tool. I've named it `commitizen-cz-cli` instead of just `commitizen`
to avoid confusion with the similarly-named, but entirely different
Python-based https://github.com/commitizen-tools/commitizen.

This implements an action with `prepare-commit-msg` hook to allow
commitizen to be injected into the normal `git commit` workflow, as well
as as a tool to expose the `cz` CLI tool.

Did my best to manually test all of this as well as I could, wasn't
really sure if you had standards for automated tests for actions/tools
but I can implement if you can point me at a doc.

Please let me know if you have questions or would like fixes made!