Add enabled version replacer callback  (#675)
Adds a replacer to patch the linter's enabled version.

This is mainly for taplo, which changes its version schema frequently
and has adopted a new release schema. In theory we could adapt this for
some of our go linters for which we can't query release versions well.

This works better as a replacer than just adapting the `preCheck`
callback because this needs to happen in setup to best propagate to the
snapshot matchers and to LUV wiring. I verified manually that the
correct version gets relayed to the LUV output.