Simplify trunk-check-pre-push-always (#672)
Change `trunk-check-pre-push-always` to not use the wrapper script. With
interactivity detection changes, we no longer need the wrapper.

This will fix the user who ran into the issue where there was a
different `trunk` in the PATH. This does not require a version
requirement update because the fix has been out since November.