Fix biome fmt error case (#772)
Fixes https://github.com/trunk-io/plugins/issues/771 for cases of
malformed config. `biome fmt` outputs an exit code of 1. For lint, we
have to have `success_codes: [0, 1]`, but for format we can just have
`success_codes: [0]` so that we better surface this error to users.

Also adds a relevant test case.