Support and migrate to eslint@9.0.0 (#735)
[Eslint@9.0.0](https://eslint.org/blog/2024/04/eslint-v9.0.0-released/)
was released and with it a major change to how eslint is configured. The
[migration
process](https://eslint.org/docs/latest/use/configure/migration-guide)
is nontrivial and may not be an option/priority for every project.
Additionally, upgrading will loudly fail unless you have accounted for
the config. Accordingly, I have done the following:

- Make `direct_configs` a property of `commands`, allowing us to version
the configs that are applicable for linter
- Users will only be upgraded/init'd/enabled at compatible versions with
their config
- Modify the `eslint` tests to handle pre- and post- eslint9
- Upgrade this repo to use an `eslint9` _flat config_
- Perform an audit and address a handful of the new diagnostics coming
from `eslint9`

This will requires a minimum version bump to support. _Most of the files
changes are lockfiles._